+++ Modules/Intent.Modules.Entities.Repositories.Api/Templates/RepositoryInterface/RepositoryInterfaceTemplatePartial.cs
@@ -1,16 +1,8 @@
-using System.Collections.Generic;
-using System.Linq;
 using Intent.Engine;
-using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
-using Intent.Modules.Common.CSharp;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
-using Intent.Modules.Common.Registrations;
-using Intent.Modules.Common.Templates;
-using Intent.Modules.Common.VisualStudio;
 using Intent.RoslynWeaver.Attributes;
-using Intent.Templates;
 
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -27,86 +19,21 @@ namespace Intent.Modules.Entities.Repositories.Api.Templates.RepositoryInterface
             : base(TemplateId, outputTarget, model)
         {
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
-                .AddUsing("System")
-                .AddUsing("System.Collections.Generic")
-                .AddUsing("System.Linq")
-                .AddUsing("System.Linq.Expressions")
-                .AddUsing("System.Threading")
-                .AddUsing("System.Threading.Tasks")
                 .AddInterface("IRepository", @interface =>
                 {
                     @interface.AddAttribute("[IntentManaged(Mode.Fully, Signature = Mode.Fully)]");
-                    @interface.AddGenericParameter("TDomain", out var tDomain)
-                        .AddGenericParameter("TPersistence", out var tPersistence)
-                        .AddMethod("void", "Add", method =>
-                        {
-                            method.AddParameter(tDomain, "entity");
-                        })
-                        .AddMethod("void", "Remove", method =>
-                        {
-                            method.AddParameter(tDomain, "entity");
-                        })
-                        .AddMethod($"Task<{tDomain}?>", "FindAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<{tDomain}?>", "FindAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter($"Func<IQueryable<{tPersistence}>, IQueryable<{tPersistence}>>", "linq")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<List<{tDomain}>>", "FindAllAsync", method =>
-                        {
-                            method.AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<List<{tDomain}>>", "FindAllAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<List<{tDomain}>>", "FindAllAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter($"Func<IQueryable<{tPersistence}>, IQueryable<{tPersistence}>>", "linq")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<IPagedResult<{tDomain}>>", "FindAllAsync", method =>
-                        {
-                            method.AddParameter("int", "pageNo")
-                                .AddParameter("int", "pageSize")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<IPagedResult<{tDomain}>>", "FindAllAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter("int", "pageNo")
-                                .AddParameter("int", "pageSize")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod($"Task<IPagedResult<{tDomain}>>", "FindAllAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter("int", "pageNo")
-                                .AddParameter("int", "pageSize")
-                                .AddParameter($"Func<IQueryable<{tPersistence}>, IQueryable<{tPersistence}>>", "linq")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod("Task<int>", "CountAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddMethod("Task<bool>", "AnyAsync", method =>
-                        {
-                            method.AddParameter($"Expression<Func<{tPersistence}, bool>>", "filterExpression")
-                                .AddParameter("CancellationToken", "cancellationToken", x => x.WithDefaultValue("default"));
-                        })
-                        .AddProperty(this.GetUnitOfWorkInterfaceName(), "UnitOfWork", prop =>
-                        {
-                            prop.ReadOnly();
-                        });
+                    @interface
+                        .AddGenericParameter("TDomain", out var tDomain, genericParameter => genericParameter.Contravariant())
+                        .AddMethod("void", "Add", method => method
+                            .AddParameter(tDomain, "entity")
+                        )
+                        .AddMethod("void", "Update", method => method
+                            .AddParameter(tDomain, "entity")
+                        )
+                        .AddMethod("void", "Remove", method => method
+                            .AddParameter(tDomain, "entity")
+                        )
+                        ;
                 });
         }