+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/Templates/TemplateExtensions.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.CustomRepository;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.CustomRepositoryInterface;
+using Intent.Modules.EntityFrameworkCore.Repositories.Templates.EfRepositoryInterface;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.PagedList;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.Repository;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.RepositoryBase;
@@ -33,6 +34,11 @@ namespace Intent.Modules.EntityFrameworkCore.Repositories.Templates
         public static string GetCustomRepositoryInterfaceName(this IIntentTemplate template, Intent.Modelers.Domain.Repositories.Api.RepositoryModel model)
         {
             return template.GetTypeName(CustomRepositoryInterfaceTemplate.TemplateId, model);
+        }
+
+        public static string GetEfRepositoryInterfaceName(this IIntentTemplate template)
+        {
+            return template.GetTypeName(EfRepositoryInterfaceTemplate.TemplateId);
         }
         public static string GetPagedListName(this IIntentTemplate template)
         {